@@ -1,15 +1,15 @@
 import React from 'react';
-import {useDispatch} from 'react-redux';
-import {BrowserRouter, Switch, Route} from 'react-router-dom';
+import {BrowserRouter, Routes, Route} from 'react-router-dom';
 
 import {getBlogPosts} from './features/blog/blogSlice';
 import './App.css';
 import BlogSingle from './routes/BlogSingle';
 import Home from './routes/Home';
 import {getExperienceList} from './features/experience/experienceSlice';
+import {useAppDispatch} from './hooks';
 
 const App = () => {
-  const dispatch = useDispatch();
+  const dispatch = useAppDispatch();
 
   React.useEffect(() => {
     dispatch(getBlogPosts(1));
@@ -18,14 +18,10 @@ const App = () => {
 
   return (
     <BrowserRouter>
-      <Switch>
-        <Route path="/blog">
-          <BlogSingle />
-        </Route>
-        <Route path="/">
-          <Home />
-        </Route>
-      </Switch>
+      <Routes>
+        <Route path="/blog" element={<BlogSingle />} />
+        <Route path="/" element={<Home />} />
+      </Routes>
     </BrowserRouter>
   );
 };
