@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import {useHistory} from 'react-router-dom';
+import {useNavigate} from 'react-router-dom';
 
 import {H3Title} from '../../styles/typography';
 
@@ -10,10 +10,10 @@ interface BlogListItemProps {
 
 const BlogList = (props: BlogListItemProps) => {
   const {blogPost} = props;
-  const history = useHistory();
+  const navigate = useNavigate();
 
   return (
-    <BlogListItemContainer onClick={() => history.push(`/blog?title=${blogPost.slug}`)}>
+    <BlogListItemContainer onClick={() => navigate(`/blog?title=${blogPost.slug}`)}>
       <H3Title>{blogPost.title.rendered}</H3Title>
       {/* eslint-disable-next-line react/no-danger */}
       <Excerpt dangerouslySetInnerHTML={{__html: blogPost.excerpt.rendered}} />
